+++ .swiftpm-locks/default/swiftImport/Package.swift
@@ -23,7 +23,9 @@ let package = Package(
     .package(path: "subpackages/_providers_kmpauth-google"),
     .package(path: "subpackages/_providers_kmpauth_facebook"),
     .package(path: "subpackages/_providers_kmpauth_google"),
-    .package(path: "subpackages/_sampleApp_shared")
+    .package(path: "subpackages/_sampleApp_shared"),
+    .package(path: "subpackages/dev_gitlive_firebase_app_3_0_0_alpha01"),
+    .package(path: "subpackages/dev_gitlive_firebase_auth_3_0_0_alpha01")
   ],
   targets: [
     .target(
@@ -39,7 +41,9 @@ let package = Package(
         .product(name: "_providers_kmpauth-google", package: "_providers_kmpauth-google"),
         .product(name: "_providers_kmpauth_facebook", package: "_providers_kmpauth_facebook"),
         .product(name: "_providers_kmpauth_google", package: "_providers_kmpauth_google"),
-        .product(name: "_sampleApp_shared", package: "_sampleApp_shared")
+        .product(name: "_sampleApp_shared", package: "_sampleApp_shared"),
+        .product(name: "dev_gitlive_firebase_app_3_0_0_alpha01", package: "dev_gitlive_firebase_app_3_0_0_alpha01"),
+        .product(name: "dev_gitlive_firebase_auth_3_0_0_alpha01", package: "dev_gitlive_firebase_auth_3_0_0_alpha01")
       ]
     )
   ]