+++ backends/firebase/kmpauth-firebase/src/iosMain/kotlin/com/mmk/kmpauth/firebase/backend/GitLivePhoneVerification.ios.kt
@@ -1,6 +1,6 @@
 package com.mmk.kmpauth.firebase.backend
 
-import cocoapods.FirebaseAuth.FIRAuthUIDelegateProtocol
+import swiftPMImport.dev.gitlive.firebase.auth.FIRAuthUIDelegateProtocol
 import com.mmk.kmpauth.core.auth.PhoneVerificationUi
 import dev.gitlive.firebase.auth.PhoneVerificationProvider
 import kotlinx.cinterop.ExperimentalForeignApi